images and centering

+++ scripts/php/feed_data_file.php
@@ -7,7 +7,11 @@
     /* get and decode existing data */
     
     $filepath = 'data/imagedata.json';
+
+    $delete = false;
     $append = true;
+    
+    $removeindex = '001';
 
     if (file_exists($filepath)) {
         $input = file_get_contents($filepath);
@@ -31,10 +35,8 @@
             $jsondata = json_encode($temparray);
             file_put_contents($filepath, $jsondata);
         }
-        else {
+        if ($delete) {
             /* remove section */
-
-            $removeindex = '001';
 
             $updated_array = delete_at_index($temparray, $removeindex);
 
@@ -48,25 +50,25 @@
         echo 'Error while loading from file: ' . $filepath;
     }
 
-function delete_at_index($array, $index) {
+    function delete_at_index($array, $index) {
 
-    $ret = [];
+        $ret = [];
 
-    for ($i = 0; $i < count($array); $i++) {
-        $image = $array[$i];
-        $image_properties = get_object_vars($image);
-        if (array_key_exists('id', $image_properties)) {
-            if ($image_properties['id'] != $index) {
-                array_push($ret, $image);
+        for ($i = 0; $i < count($array); $i++) {
+            $image = $array[$i];
+            $image_properties = get_object_vars($image);
+            if (array_key_exists('id', $image_properties)) {
+                if ($image_properties['id'] != $index) {
+                    array_push($ret, $image);
+                }
+            }
+            else {
+                echo "Encountered a problem while reading data : unefined id\n";
             }
         }
-        else {
-            echo "Encountered a problem while reading data : unefined id\n";
-        }
-    }
 
-    return $ret;
-}
+        return $ret;
+    }
 
 
 ?>
